to best attempt to handle up navigation we include any common extras we know about
